Update MessageValidator.php
At the expense of having to explain to my wife why I have a bald spot on the crown of my head, we figured out that this small issue is what ailed us all afternoon. Messages do not validate with the existing method had to change to check for empty().