Disallow null-byte-sending encoders for TCP
Since GELF via TCP uses \0 to separate multiple messages over the same
tcp connection, any encoder which may emit \0 for a message is not
allowed with the TcpTransport as it can break. This commit introduces
a new interface which is only implemented by the uncompressed JSON
encoder. This is checked at runtime by the TcpTransport to make sure
it doesn't break